refactored utils.h
